==> 01_timetable_schedule.ipynb <==
# **Getting the timetable schedule from APIs**

import requests
import json
import pandas
from pprint import pprint
import os

file_key = open('api_key.txt', encoding='utf-8')

key = [key.strip() for key in file_key]

api_key = key[0]

headers = {
  'Content-Type': 'application/json; charset=utf-8',
  'x-access-token': api_key
}

trip_id = ['331_243_200302', '331_257_200302', '333_677_200302',
       '333_1072_200302', '333_557_200302', '331_241_200302',
       '333_1064_200302', '333_1115_200302', '331_260_200302',
       '333_1012_200302', '333_1052_200302', '333_1094_200302',
       '333_715_200302', '333_751_200302', '333_651_200307',
       '333_1026_200307', '333_653_200307', '333_671_200307',
       '333_1010_200307', '333_1037_200307', '333_636_200307',
       '333_1014_200307', '333_1047_200307', '333_655_200307',
       '333_1024_200307', '333_673_200307', '333_1012_200307']

schedules = []
for tripid in trip_id:
    api_key = "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJlbWFpbCI6ImphbmR1bGFocnViYUBnbWFpbC5jb20iLCJpZCI6MjY0LCJuYW1lIjpudWxsLCJzdXJuYW1lIjpudWxsLCJpYXQiOjE1ODc2NDg5ODksImV4cCI6MTE1ODc2NDg5ODksImlzcyI6ImdvbGVtaW8iLCJqdGkiOiJiMGQzNGQzMC03ODNlLTQ0MWQtOTQwMC04MGMwNmZmYTQ3NDkifQ.j0q7LKzycT3zmVLxxkPlWKIYCkl_PYfjgIraFtRQ1YA"
    headers = {
    'Content-Type': 'application/json; charset=utf-8',
    'x-access-token': api_key
    }

    response1 = requests.get(f'https://api.golemio.cz/v2/gtfs/trips/{tripid}?includeStops=true&includeStopTimes=true', headers = headers)
    #print(response1)
    
    if response1.status_code == 200:
        info_all = json.loads(response1.text)
       
        final = pandas.json_normalize(info_all['stop_times'])
        #print(final)
        final.drop(inplace=True,columns=['arrival_time_seconds','departure_time_seconds','drop_off_type','pickup_type','stop_headsign',
                                 'stop.properties.stop_code','stop.properties.stop_desc','stop.properties.stop_timezone','stop.properties.stop_url',
                                 'stop.properties.wheelchair_boarding','stop.properties.zone_id','stop.type','stop.geometry.type','stop.properties.level_id','stop.properties.location_type',
                                 'stop.properties.parent_station','stop.properties.platform_code','stop.properties.stop_id','timepoint'])
        final.columns = ['ARRIVAL_TIME','DEPARTURE_TIME', 'SHAPE_DIST_TRAVELED', 'STOP_ID','STOP_SEQUENCE', 'TRIP_ID', 'GEOMETRY_COORDINATES', 'STOP_NAME']
        schedules.append(final)


result = pandas.concat(schedules)
result.to_csv('timetable.csv', encoding="utf-8", index=False)

timetable = pandas.read_csv('timetable.csv', encoding='utf-8', low_memory=False)

timetable
