==> 02_geomap.ipynb <==
# **Geo map of the bus route**

import json
import ipyleaflet as ipy 
from ipyleaflet import (
    Map,
    Marker,
    Icon,
    TileLayer, ImageOverlay,
    Polyline, Polygon, Rectangle, Circle, CircleMarker,
    GeoJSON,
    DrawControl
)

import matplotlib as mpl
import matplotlib.cm
import matplotlib.colors
import numpy as np

def n_colors(n, colormap=mpl.cm.Blues):
    data = np.linspace(0.0,1.0,n)
    c = [mpl.colors.rgb2hex(d[0:3]) for d in colormap(data)]
    return c

def data_to_colors(data, colormap=mpl.cm.Blues):
    c = [mpl.colors.rgb2hex(d[0:3]) for d in colormap(mpl.colors.Normalize()(data))]
    return c

with open('bus_stops_333_Brezany.json') as f:
    data = json.load(f)
#data

geo_json_stops = [x['stop'] for x in data]
#geo_json_stops

geo_json = { "type": "FeatureCollection", "features": geo_json_stops}

n_features = len(geo_json['features'])
colors = n_colors(n_features)
print(n_features)

#colors

for feature, color in zip(geo_json['features'], colors):
    feature['properties']['style'] = {'color':color, 'weight': 1, 'fillColor':color, 'fillOpacity':0.5, 'opacity':0.5}

g = GeoJSON(data=geo_json)

center = [50.025753, 14.455383]
zoom = 12.5
m = Map(center = center, zoom = zoom)

for feature in geo_json['features']:
    icon = Icon(icon_url='bus_png.png', icon_size=[30, 50], icon_anchor=[12,50]) 
    mark = Marker(location=[feature['geometry']['coordinates'][1],feature['geometry']['coordinates'][0]], icon=icon)
    m.add_layer(mark)


m
