==> 03_stops_vs_positions.ipynb <==
# **Creating pre-final dataset**

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import json
import csv

### **Loading of all data**

data_Honza = pd.read_csv('out.c-cleaning.ALL_HONZA_NEW.csv', encoding='utf-8', low_memory=False)
data_Honza_wkend = pd.read_csv('out.c-cleaning.ALL_HONZA_NEW_WKEND.csv', encoding='utf-8', low_memory=False)
data_keboola = pd.read_csv('out.c-cleaning.ALL_POSITIONS_NEW.csv', encoding='utf-8', low_memory=False)
data_keboola_wkend = pd.read_csv('out.clean_VIKEND_END.csv', encoding='utf-8', low_memory=False)

data = pd.concat([data_Honza,data_Honza_wkend,data_keboola,data_keboola_wkend], ignore_index = True)
data = data.dropna(axis=0,subset=['DELAY'],how='any')

### **Adding columns for route name and day of week** 

data['ROUTE_NAME'] = data['TRIP_ID'].str[:3]

data['DATE'] = data['DATE'] + " 2020"

data['DATE'] = pd.to_datetime(data['DATE'])

data['DAY_OF_WEEK'] = data['DATE'].dt.day_name()

data['DAY_OF_WEEK_NUM'] = data['DATE'].dt.dayofweek

### **We want route 331,333 +  in the morning traffic + towards the subway station**

morn_331_333 = data.loc[(data['HEADSIGN'] == 'Praha,Kačerov') 
                     & (data['START_TIME'] < '12:00:00') 
                     & ((data['ROUTE_NAME'] == '333') | (data['ROUTE_NAME'] == '331'))
                     , ['UNICORN', 'TRIP_ID', 'START_TIME', 'HEADSIGN', 'DATE',
                        'DELAY', 'DELAY_STOP_ARRIVAL', 'DELAY_STOP_DEPARTURE',
                        'ORIGIN_TIMESTAMP', 'SHAPE_DIST_TRAVELED', 'SPEED', 'LATITUDE',
                        'LONGITUDE','ROUTE_NAME','DAY_OF_WEEK']]

#morn_331_333.head()

# count of unique trip_id
len(morn_331_333['TRIP_ID'].unique())

# count of uniques unicorns z df data (allpositions)
unicorn_list = morn_331_333['UNICORN'].unique().tolist()
len(unicorn_list)

### **Dataframe with timetable schedule for all uniques trip_id (jizdni_rad.ipynb)**

info_stops_clean = pd.read_csv('timetable.csv', encoding='utf-8', low_memory=False)

info_stops_clean

# selection of columns we want
stop_dist = info_stops_clean.loc[:,['TRIP_ID','SHAPE_DIST_TRAVELED','STOP_ID','STOP_SEQUENCE','STOP_NAME','ARRIVAL_TIME']]
# stop_dist

## **Empty DF and appending all the data we want -> merging bus and stop info**

# new DF
final_tab = pd.DataFrame()

index = 0
final_tab = pd.DataFrame()

for uni in unicorn_list:
    #to see the progress
    print(round(index/len(unicorn_list)*100), '%')
    index+=1
    
    #smaller dataframe based on selected uni
    unicorn_find = morn_331_333[morn_331_333['UNICORN'] == uni]
    
    #columns I want from the bus (all positions) dataframe
    bus_dist = unicorn_find.loc[:,['TRIP_ID','SHAPE_DIST_TRAVELED','DELAY','DATE','DAY_OF_WEEK',
                                   'START_TIME','HEADSIGN','ROUTE_NAME','ORIGIN_TIMESTAMP']]
    
    #these values are repeating for one unicorn so I want just the 1 to tell me the information
    gtfs_trip_id = bus_dist['TRIP_ID'].tolist()[0]
    trip_date = bus_dist['DATE'].tolist()[0]
    trip_day = bus_dist['DAY_OF_WEEK'].tolist()[0]
    trip_start_time = bus_dist['START_TIME'].tolist()[0]
    trip_headsign = bus_dist['HEADSIGN'].tolist()[0]
    route_name = bus_dist['ROUTE_NAME'].tolist()[0]    
    
    #for every bus stop in timetable for GTFS_TRIP_ID
    for i, stop in stop_dist[stop_dist['TRIP_ID'] == gtfs_trip_id].iterrows():
        
        #I am looking for the closest position(value), I am gonna set the first value as really far away
        min_distance_so_far = 100000
        #and delay to 0 
        stop_delay = 0
        
        for j, position in bus_dist.iterrows():
            
            # absolute value of the distance of stop - position
            actual_distance = abs(stop['SHAPE_DIST_TRAVELED'] - position['SHAPE_DIST_TRAVELED'])

            # if this value is closer than 100000 or closer than the last one found
            if (min_distance_so_far > actual_distance):
                #save it
                min_distance_so_far = actual_distance
                #with delay and origin timestamp
                stop_delay = position['DELAY']
                origin_timestamp = position['ORIGIN_TIMESTAMP'].split(' ')[4]
            
        # now for the bus stop I checked all the positions and return only the closest one
        #look of the final DF with columns and values I want
        df = pd.DataFrame(
            {
                "TRIP_ID": [gtfs_trip_id],
                "ROUTE" : [route_name],
                "DAY_OF_WEEK": [trip_day],
                "DATE": [trip_date],
                "UNICORN": [uni],
                "START_TIME" : [trip_start_time],
                "HEADSIGN" : [trip_headsign],
                "STOP_ID": [stop['STOP_ID']],
                "STOP_SEQUENCE": [stop['STOP_SEQUENCE']],
                "STOP_NAME" : [stop['STOP_NAME']],
                "SHAPE_DIST_TRAVELED": [stop['SHAPE_DIST_TRAVELED']],
                "DELAY": [stop_delay],
                "ORIGIN_TIMESTAMP": [origin_timestamp],
                "TIMETABLE_ARRIVAL_TIME":[stop['ARRIVAL_TIME']]
            }
        )
        # append to the big DF
        final_tab = final_tab.append(df, ignore_index = True)

#just check the count of trip_ids
len(final_tab['UNICORN'].unique())

#final_tab

final_tab['TRIP_ID'].unique()

Unique trip_id --> unification of the number of bus stops -> STOP_SEQUENCE from Dolní Břežany,Náměstí

- 333_655_200307 >= 2
- 333_1012_200307 >= 14 
- 333_655_200307 Dolní Břežany,Náměstí
- 333_1047_200307(Zvole) >= 7 
- 333_636_200307 Dolní Břežany,Náměstí
- 333_1037_200307(Březová-Oleško,Oleško) >= 14 
- 333_1010_200307(Březová-Oleško,Oleško) >= 19   
- 333_653_200307 Dolní Břežany,Náměstí	
- 333_1026_200307(Zvole) >= 14
- 333_651_200307 Dolní Břežany,Náměstí	
- 333_1115_200302 (Zvole) >=12
- 333_1094_200302	(Březová-Oleško,Oleško) >= 14
- 333_1012_200302(Březová-Oleško,Oleško) >= 14
- 333_751_200302 (Dolní Břežany,Obecní úřad) >= 2
- 333_677_200302 (Dolní Břežany,Obecní úřad) >= 2
- 333_1052_200302(Zvole) >= 12
- 333_1072_200302(Zvole,Nová Zvole) >= 14
- 333_715_200302 (Dolní Břežany,Obecní úřad) >= 2
- 333_1064_200302 (Březová-Oleško,Oleško) >= 14
- 333_557_200302 (Dolní Břežany,Zálepy) >= 7
- 333_1014_200307 (Zvole) >= 14
- 333_671_200307 Dolní Břežany,Náměstí
- 333_1024_200307 (Zvole) >= 7
- 333_673_200307 Dolní Břežany,Náměstí

- 331_241_200302 (Jílové u Prahy,Náměstí) >= 14	
- 331_243_200302 (Jílové u Prahy,Náměstí) >= 14	
- 331_257_200302 (Okrouhlo,Zahořany) >= 11
- 331_260_200302 (Zvole) >= 8

# I will just choose the ones they have in common and they are closet to the city
final_clean =   final_tab[
((final_tab['TRIP_ID'] == '333_677_200302') & (final_tab['STOP_SEQUENCE'] >= 2)) 
| ((final_tab['TRIP_ID'] == '333_715_200302') & (final_tab['STOP_SEQUENCE'] >= 2))
| ((final_tab['TRIP_ID'] == '333_751_200302') & (final_tab['STOP_SEQUENCE'] >= 2))
| ((final_tab['TRIP_ID'] == '333_1012_200307') & (final_tab['STOP_SEQUENCE'] >= 14))
| ((final_tab['TRIP_ID'] == '333_1047_200307') & (final_tab['STOP_SEQUENCE'] >= 7))
| ((final_tab['TRIP_ID'] == '333_1037_200307') & (final_tab['STOP_SEQUENCE'] >= 14))
| ((final_tab['TRIP_ID'] == '333_1010_200307') & (final_tab['STOP_SEQUENCE'] >= 19))
| ((final_tab['TRIP_ID'] == '333_1026_200307') & (final_tab['STOP_SEQUENCE'] >= 14))
| ((final_tab['TRIP_ID'] == '333_1115_200302') & (final_tab['STOP_SEQUENCE'] >= 12))
| ((final_tab['TRIP_ID'] == '333_1094_200302') & (final_tab['STOP_SEQUENCE'] >= 14))
| ((final_tab['TRIP_ID'] == '333_1012_200302') & (final_tab['STOP_SEQUENCE'] >= 14))
| ((final_tab['TRIP_ID'] == '333_1052_200302') & (final_tab['STOP_SEQUENCE'] >= 12))
| ((final_tab['TRIP_ID'] == '333_1072_200302') & (final_tab['STOP_SEQUENCE'] >= 14))
| ((final_tab['TRIP_ID'] == '333_1064_200302') & (final_tab['STOP_SEQUENCE'] >= 14))
| ((final_tab['TRIP_ID'] == '333_557_200302') & (final_tab['STOP_SEQUENCE'] >= 7))
| ((final_tab['TRIP_ID'] == '331_241_200302') & (final_tab['STOP_SEQUENCE'] >= 14))
| ((final_tab['TRIP_ID'] == '331_243_200302') & (final_tab['STOP_SEQUENCE'] >= 14))
| ((final_tab['TRIP_ID'] == '331_257_200302') & (final_tab['STOP_SEQUENCE'] >= 11))
| ((final_tab['TRIP_ID'] == '331_260_200302') & (final_tab['STOP_SEQUENCE'] >= 8))
| (final_tab['TRIP_ID'] == '333_636_200307')
| (final_tab['TRIP_ID'] == '333_655_200307')
| (final_tab['TRIP_ID'] == '333_653_200307')
| (final_tab['TRIP_ID'] == '333_651_200307')
| (final_tab['TRIP_ID'] == '333_671_200307')
| ((final_tab['TRIP_ID'] == '333_1014_200307') & (final_tab['STOP_SEQUENCE'] >= 14))
| ((final_tab['TRIP_ID'] == '333_1024_200307') & (final_tab['STOP_SEQUENCE'] >= 7))
| (final_tab['TRIP_ID'] == '333_673_200307')
]

final_clean

#final check
#final_clean[(final_clean['STOP_NAME'] == 'Dolní Břežany,Náměstí')]
len(final_clean.loc[:,'UNICORN'].unique())

final_clean.to_csv('stops_vs_positions_clean.csv', encoding="utf-8", index=False)
