==> 04_diff.ipynb <==
# **Calculation of the delay difference and previous delay**

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import json
import csv

stops_vs_positions_diff = pd.read_csv('stops_vs_positions_clean.csv', encoding="utf-8", low_memory=False, index_col=False)
stops_vs_positions_diff.head()

unicorn_list = stops_vs_positions_diff['UNICORN'].unique().tolist()
len(unicorn_list)

for uni in unicorn_list:
    unicorn_find = stops_vs_positions_diff[stops_vs_positions_diff['UNICORN'] == uni]
    stops_vs_positions_diff['DIFF_DELAY'] = stops_vs_positions_diff["DELAY"].diff(periods=1)
    stops_vs_positions_diff.loc[stops_vs_positions_diff['STOP_ID'] == 'U1502Z2', ['DIFF_DELAY']] = 0
    stops_vs_positions_diff['PREV_DELAY'] = stops_vs_positions_diff["DELAY"].shift(1)

stops_vs_positions_diff.tail()

sekv_df = []
for uni in unicorn_list:
    
    unicorn_find = stops_vs_positions_diff[stops_vs_positions_diff['UNICORN'] == uni]

    sekv = unicorn_find.assign(SEQUENCE = [1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15] )
    sekv_df.append(sekv)

final = pd.concat(sekv_df)

final.to_csv('final.csv',encoding="utf-8", index=False)

final
